requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.62
pydantic==2.5.0
streamlit==1.28.1
gitpython==3.1.40
python-decouple==3.8
pymysql==1.1.0
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
openai==1.3.7
requests==2.31.0
rank-bm25==0.2.2
cyclonedx-python-lib==6.4.4
jsonschema==4.20.0
tiktoken==0.5.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
plotly==5.17.0
pandas==2.1.3
pyyaml==6.0.1
cryptography==41.0.7
google-generativeai==0.3.2
license-expression==30.4.4
bcc==0.1.7  # Optional: Required only for eBPF runtime tracing
psutil==5.9.6